Use `find_all` to get all substitutions
I had been using `find` function in finding the rows that corresponded
to a substitution which led just recording one substitution if more
than one substitutions were being made at the same time. Changed the
call to `find_all` and looping over all the rows to now record all the
substitutions instead. As can be seen in the data files, this has led
to the inclusion of several instances where we were missing 1 or even
2 other substitutions.