fix case where starting squad has too many players per team
e.g. when a player moves club during the season you can end up with 4 players from the same club.

Also update creating the initial squad to use player attributes (i.e. team) from the current gameweek, not from when the transaction was made.

Note that this makes the optimisation slower. I think this is because it takes longer to find a valid transfer, so it should hopefully be the same/similar as before if starting with a valid squad, but tbc.